oh-tab-agno: extract LlmProfilesView form state
